// web-2/src/api/generated/model/userResponseDto.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * api
 * OpenAPI spec version: v1
 */

export interface UserResponseDto {
  id?: string;
  email?: string;
  firstName?: string;
  lastName?: string;
  isActive?: boolean;
  isEmailVerified?: boolean;
  createdAt?: string;
  /** @nullable */
  lastLoginAt?: string | null;
  lastSeen?: string;
}